Use express router and install sequelize
